requirements.txt
fastapi
uvicorn
python-multipart
transformers
torch
scipy
python-doctr[torch]
